+++ src/main/kotlin/icu/windea/pls/config/cwt/config/CwtAliasConfig.kt
@@ -2,9 +2,15 @@ package icu.windea.pls.config.cwt.config
 
 import com.intellij.psi.*
 import icu.windea.pls.*
+import icu.windea.pls.config.cwt.*
 import icu.windea.pls.config.cwt.expression.*
 import icu.windea.pls.cwt.psi.*
 
+/**
+ *
+ * @property supportedScopes (option) scope/scopes: string | string[]
+ * @property supportedScopeNames 所有支持的作用域的名字。
+ */
 data class CwtAliasConfig(
 	override val pointer: SmartPsiElementPointer<CwtProperty>,
 	override val info: CwtConfigInfo,
@@ -17,12 +23,15 @@ data class CwtAliasConfig(
 	
 	//TODO check
 	
-	val supportedScopes by lazy { resolveSupportedScopes() }
-	val supportedScopesText by lazy { supportedScopes?.joinToString(" ", "{ ", " }") }
-	
-	private fun resolveSupportedScopes(): Set<String>? {
-		val options = config.options ?: return emptySet()
-		val option = options.find { it.key == "scope" || it.key == "scopes" } ?: return null
-		return option.stringValue?.let { setOf(it) } ?: option.optionValues?.mapNotNullTo(mutableSetOf()) { it.stringValue } ?: emptySet()
+	val supportedScopes = config.options
+		?.find { o -> o.key == "scope" || o.key == "scopes" }
+		?.let { o -> o.stringValue?.let { setOf(it) } ?: o.optionValues?.mapNotNullTo(mutableSetOf()) { it.stringValue } }
+	val supportAnyScope = supportedScopes.isNullOrEmpty() || supportedScopes.singleOrNull() == "any"
+	val supportedScopeNames by lazy {
+		if(supportAnyScope) {
+			setOf("Any")
+		} else {
+			supportedScopes?.mapTo(mutableSetOf()) { CwtConfigHandler.getScopeName(it, info.configGroup) }.orEmpty()
+		}
 	}
 }